Fork 'tableflip' to coordinate over a unix socket
This removes the parent/child process forking stuff in favor of using a
well-known path for the parent and child to discover each other.

Filesystem locking and unix sockets are used to coordinate file
descriptor passing in this new world.

This commit includes the base of the rework, though there is still
significant work to do.