Remove 'env'
The whole 'env' thing is kinda silly imo. The better way to do this
unit-testing-stubbing thing is to have e.g. an 'os' interface with a
mock and real implementation. There are many examples of this floating
around online.

For now, I've removed all unit testing so we can just drop it entirely
and simplify things.