Remove 'singleton' requirement from upgrader
It doesn't really make much sense that upgrader needs to be a singleton.
The 'exitFd' thing might matter, but in reality the upgrader will
usually be constructed in main and kept in scope so it can't be GCd
anyway.
I also suspect the exit detection will be reworked in a coming commit ;)