fds: rework to be id based and allow removal
This does a complete rework of the file descriptor store. It is not
backwards compatible with previous versions, and since there have been
no releases / users of tableroll yet, it's a hard breaking change with
no effort made to provide any migration path.

This also changes tableroll's "Fds" api to be entirely based on unique
identifiers for each listener.

Every way to add a fd to the store now requires the use of an ID, and
only the ID is used for retreiving said fds.

The idea of "inherited and used" file descriptors has also vanished. If
a program wishes to remove a file descriptor before, it would simply
avoid using it (e.g. a program that used to listen on :8080 and :8081
would simply no longer listen on :8081 after an upgrade, and that fd
would vanish).
After this change, all file descriptors are considered necessary if not
explicitly removed. To remove a now-unused listener after upgrade, the
'Remove' api must be called with the id after the upgrader is
constructed.

In addition to the focus on ids, this also attempts to allow better
locking support in the api through serializing all file descriptor
changes, and through encouraging users to create new listeners by
passing in a function which will have its execution guarded by a lock.

This paves the way for a world in which fds can be more easily mutated
over time, and where the draining state of an upgrader prevents new
listeners from being constructed.