Only transfer one fd at a time
The previous code would pass all file descriptors as auxiliary data in a
single cmsg. This, it turns out, very easily hits limits on the size of
cmsg aux data. In practice, this happened at around 300 to 500 listening
sockets.

This changes the code to only pass one at a time, adds a test that
failed before, and for good measure switches to a different upstream
library for doing the transfer.