fix compilation on Darwin
This package does not compile on Darwin because a file in
runc/libcontainer/utils contains a constant that does not exist on
Darwin builds.

The code out of that package that we need has no problems compiling on
Darwin however so pull it into a shim and load it from there.

If opencontainers/runc#2048 is resolved maybe we can re-import the
other package.