Remove usage of the os pid entirely
It doesn't play nicely with pid namespaces, and if the user wants to use
the pid, they can pass it in.

This does remove our ability to check original owner liveness, other
than via file locks being held or not, but if the caller's using us
correctly, then the "id lock files" should be sufficient for our
purpose.

Fixes #21