Don't call '.Fd' while sending file descriptors
The '*os.File.Fd' call, as a side effect, puts a file into blocking mode.

We don't want this because we've dup'd file descriptors (all of which
share the same file description), and setting an fd to blocking mode
impacts all of the fds sharing the same file description.

As a result, this 'Fd' call could cause a deadlock, which the test in
the previous commit exhibits.

This fixes that by replacing the 'RecvFd' and 'SendFd' api we were using
with an internal one that intentionally keeps a reference to the fd
around so we can skip calling '*os.File.Fd()'

The actual wire format of the sendmsg/recvmsg calls is identical, so
this remains backwards compatible, we just store the sent/recv'd fd
slightly differently.

With this change, the test in the previous commit now passes.