Drop v1->v2 upgrade test
This was causing us to have an import of an old tableroll which the
modern go tooling did not like the go mod of.

Specifically, I was getting this error:

```
go: github.com/ngrok-oss/tableroll/v2@v2.0.1 requires
	github.com/euank/tableroll@v1.0.1-0.20190509205925-1c7855724129 requires
	github.com/opencontainers/runc@v0.0.0-00010101000000-000000000000: invalid version: unknown revision 000000000000
```

I think it's fine to drop this test now. It's served its purpose, and we
don't really expect anyone to still be on v1 and need to upgrade.

More importantly, this makes `go get`ing this code work with newer code